added simple (single-RHS) versions of loggers and stopping criteria to reference/omp
