+++ reference/log/batch_logger.hpp
@@ -44,6 +44,48 @@ namespace batch_log {
 
 
 /**
+ * Logs the final residual and iteration count for a batch solver.
+ *
+ * Specialized for a single RHS.
+ */
+template <typename RealType>
+class SimpleFinalLogger final {
+public:
+    using real_type = RealType;
+
+    /**
+     * Sets pre-allocated storage for logging.
+     *
+     * @param batch_residuals  Array of residuals norms of size
+     *                         num_batches x num_rhs. Used as row major.
+     * @param batch_iters  Array of final iteration counts for each
+     *                     linear system and each RHS in the batch.
+     */
+    SimpleFinalLogger(real_type* const batch_residuals, int* const batch_iters)
+        : final_residuals_{batch_residuals}, final_iters_{batch_iters}
+    {}
+
+    /**
+     * Logs the iteration count and residual norm.
+     *
+     * @param batch_idx  The index of linear system in the batch to log.
+     * @param iter  The current iteration count (0-based).
+     * @param res_norm  Norm of current residual
+     */
+    void log_iteration(const size_type batch_idx, const int iter,
+                       const real_type res_norm)
+    {
+        final_iters_[batch_idx] = iter;
+        final_residuals_[batch_idx] = res_norm;
+    }
+
+private:
+    real_type* const final_residuals_;
+    int* const final_iters_;
+};
+
+
+/**
  * Simple logger for final residuals and iteration counts of all
  * linear systems in a batch.
  */
@@ -63,7 +105,7 @@ public:
      *                     linear system and each RHS in the batch.
      */
     FinalLogger(const int num_rhs, const int max_iters,
-                real_type *const batch_residuals, int *const batch_iters)
+                real_type* const batch_residuals, int* const batch_iters)
         : nrhs_{num_rhs},
           max_iters_{max_iters},
           final_residuals_{batch_residuals},
@@ -89,7 +131,7 @@ public:
      * @param converged  Bitset representing convergence state for each RHS.
      */
     void log_iteration(const size_type batch_idx, const int iter,
-                       const real_type *const res_norm, const uint32 converged)
+                       const real_type* const res_norm, const uint32 converged)
     {
         if (iter == 0) {
             init_converged_ = 0 - (1 << nrhs_);
@@ -116,8 +158,8 @@ public:
 private:
     const int nrhs_;
     const int max_iters_;
-    real_type *const final_residuals_;
-    int *const final_iters_;
+    real_type* const final_residuals_;
+    int* const final_iters_;
     uint32 init_converged_;
 };
